pyproject.toml
[tool.poetry]
name = "openai2mqtt"
version = "0.1.0"
description = "OpenAI API to MQTT Bridge"
authors = ["Louis King <jinglemansweep@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
dynaconf = "^3.1.12"
python = "^3.9"
python-dotenv = "^0.21.1"
paho-mqtt = "^1.6.1"
pyyaml = "^6.0"
types-paho-mqtt = "^1.6.0.5"
types-pyyaml = "^6.0.12.12"
pydantic = "^2.5.3"
openai = "^1.63.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
gitlint = "^0.19.1"
mypy = "^1.0.0"
pre-commit = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
